\name{lulu-package}
\alias{lulu-package}
\alias{lulu}
\docType{package}
\title{
\packageTitle{lulu}
}
\description{
\packageDescription{lulu}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lulu}
\packageIndices{lulu}
An overview of how to use the package, including the most important
functions
}
\author{
\packageAuthor{lulu}

Maintainer: \packageMaintainer{lulu}
}
\references{
Frøslev, T. G., Kjøller, R., Bruun, H. H., Ejrnæs, R., Brunbjerg, A. K., Pietroni, C., & Hansen, A. J. (2017). Algorithm for post-clustering curation of DNA amplicon data yields reliable biodiversity estimates. Nature Communications, 8(1), 1188.
}
\keyword{ package pene}
\seealso{
Optional links to other man pages, e.g.
}
\examples{
You can run lulu as:
    curated_table <- lulu("otufilename", "matchfilename", )
}
