% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{lulu}
\alias{lulu}
\title{Post Clustering Curation of Amplicon Data.}
\usage{
lulu(otutable, matchlist, abundance_threshold = "min", minimum_match = 84,
  minimum_relative_cooccurence = 0.95)
}
\arguments{
\item{otutable}{a data.frame with with an OTU table that has sites/samples as
columns and OTUs (unique OTU id's) as rows, and observations as read
counts.}

\item{matchlist}{a data.frame containing three columns: (1) OTU id of
potential child, (2) OTU id of potential parent, (3) match - \% identity
between the sequences of the potential parent and potential child OTUs.
\strong{NB: The matchlist is the product of a mapping of OTU sequences against each other. This is
currently carried out by an external script in e.g. Blastn or VSEARCH, prior to running lulu!}}

\item{abundance_threshold}{sets whether a potential error must have lower
abundance than the parent in all samples \code{min} (default), or if an error
just needs to have lower abundance on average \code{avg}. Choosing lower
abundance on average over globally lower abundance will greatly increase
the number of designated errors. This option was introduced to make it
possible to account for non-sufficiently clustered intraspecific variation,
but is not generally recommended, as it will also increase the potential of
cluster well-separated, but co-occuring, sequence similar species.}

\item{minimum_match}{minimum threshold of sequence similarity
for considering any OTU as an error of another can be set (default 84\%).}

\item{minimum_relative_cooccurence}{minimum co-occurrence rate – i.e. the
lower rate of occurrence of the potential error explained by co-occurrence
with the potential parent for considering error state.}
}
\value{
Function \code{lulu} returns a list of results based on the input OTU
  table and match list.
  \enumerate{
  \item \code{curated_table} - a curated
  OTU table with daughters merged with their matching parents.
  \item \code{curated_count} - number of curated (parent) OTUs.
  \item \code{curated_otus} - ids of the OTUs that were accepted as valid OTUs.
  \item \code{discarded_count} - number of discarded (merged with parent) OTUs.
  \item \code{discarded_otus} - ids of the OTUs that were identified as
  errors (daughters) and merged with respective parents.
  \item \code{runtime} - time used by the script.
  \item \code{minimum_match} - the id threshold
  (minimum match \% between parent and daughter) for evaluating co-occurence (set
  by user).
  \item \code{minimum_relative_cooccurence} - minimum ratio of
  daughter-occurences explained by co-occurence with parent (set by user).
  \item \code{otu_map} - information of which daughters were mapped to which
  parents.
  \item \code{original_table} - original OTU table. }
}
\description{
This algorithm \code{lulu} consumes an OTU table and a matchlist, and
  evaluates cooccurence of 'daughters' (potential analytical artefacts) and
  their 'parents' (≈ real biological species/OTUs). The algorithm requires an
  OTU table (species/site matrix), and a match list. The OTU table can be
  made with various r-packages (e.g. \code{DADA2}) or external pipelines (\code{VSEARCH, USEARCH, QIIME}, etc.), and the
  match-list can be made with external bioinformatic tools like \code{VSEARCH, USEARCH, BLASTN} or another algorithm
  for pair-wise sequence matching.
}
\details{
\strong{The matchlist is the product of a mapping of OTU sequences against each other. This is
  currently carried out by an external script in e.g. \code{BLASTN} or \code{VSEARCH}, prior to running \code{lulu}!}
\describe{
\item{Producing the match list requires a file with all
the OTU sequences (centroids) - e.g. \code{OTUcentroids.fasta}. The matchlist can
be produced by mapping all OTUs against each other with an external algorithm
like \code{VSEARCH} or \code{BLASTN} }
\enumerate{
\item \code{VSEARCH} - in VSEARCH a
matchlist can be produced e.g. with the following command:\cr \code{vsearch
--usearch_global OTUcentroids.fasta --db OTUcentroids.fasta --strand plus
--self --id .80 --iddef 1 --userout matchlist.txt --userfields
query+target+id --maxaccepts 0 --query_cov .9 --maxhits 10}
\item \code{BLASTN} - in BlastN a matchlist can be produces e.g. with the following
commands. First we produce a blast-database from the fasta file:\cr
\code{makeblastdb -in OTUcentroids.fasta -parse_seqids -dbtype nucl}\cr then we
match the centroids against that database:\cr \code{blastn -db OTUcentoids.fasta
-num_threads 10 -outfmt'6 qseqid sseqid pident' -out matchlist.txt
-qcov_hsp_perc .90 -perc_identity .84 -query OTUcentroids.fasta}
}
}
}
\examples{
lulu(my_table, my_matchlist)
lulu(my_table, my_matchlist, abundance_threshold = "avg", minimum_match = 0.8, minimum_relative_cooccurence = 0.9)
}
\author{
Tobias Guldberg Frøslev
}
